Add files via upload
This one is my latest and by far the best website that I've created, I made this literally everything that I've learnt so far in my web development learning journey I've used CSS grid, Flexbox and different transitions and made it responsive for most of the devices and I've also used little bit JavaScript as well so that I can make few things via function such as a carousel. 
Apart from this, this website is based on an idea as well which is its a blogging website which is made to express all the intricacies in a novel for instances there character build-up, different conditions that a person has go through, their decision making and showing their cultures as well in different eras too. So in my opinion novels have a lot to offer it can make you a better thinker and a better decision maker since when you read similar events from your life that happened in someone else's life you go into a deep introspection which is very good for self actualization and better decision making.